# Reduce the memory allocation by refactoring explodePath method