# Overview
I felt like I could reduce the memory allocation. 🤔 

The Search method seems to be where the bottleneck is most hidden, but the Search method's processing is a little complicated, so first improve explodePath, which seems to be the second bottleneck and has simple processing.

# Changes
Run commands, and compare results.
```sh
go test -bench=. -cpu=1 -benchmem
go test -bench . -memprofile mem.out && go tool pprof -http=":8889" mem.out
```

Before
```sh
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         5072353               240.1 ns/op           128 B/op          4 allocs/op
BenchmarkStatic5         2491546               490.0 ns/op           384 B/op          6 allocs/op
BenchmarkStatic10        1653658               729.6 ns/op           720 B/op          7 allocs/op
BenchmarkWildCard1       1602606               747.3 ns/op           456 B/op          9 allocs/op
BenchmarkWildCard5        435784              2716 ns/op            1016 B/op         23 allocs/op
BenchmarkWildCard10       246729              5033 ns/op            1680 B/op         35 allocs/op
BenchmarkRegexp1         1647258               733.2 ns/op           456 B/op          9 allocs/op
BenchmarkRegexp5          456652              2641 ns/op            1016 B/op         23 allocs/op
BenchmarkRegexp10         251998              4780 ns/op            1680 B/op         35 allocs/op
PASS
ok      github.com/bmf-san/goblin       14.304s
```

![スクリーンショット 2022-11-03 1 00 43](https://user-images.githubusercontent.com/13291041/199539662-8f4d7b2d-267d-4c7e-87da-e01e243c7c6d.png)


After
```sh
goos: darwin
goarch: amd64
pkg: github.com/bmf-san/goblin
cpu: VirtualApple @ 2.50GHz
BenchmarkStatic1         8243053               145.8 ns/op            64 B/op          2 allocs/op
BenchmarkStatic5         4207813               282.6 ns/op           128 B/op          2 allocs/op
BenchmarkStatic10        2626821               451.4 ns/op           208 B/op          2 allocs/op
BenchmarkWildCard1       1766892               685.1 ns/op           424 B/op          8 allocs/op
BenchmarkWildCard5        456736              2531 ns/op             760 B/op         19 allocs/op
BenchmarkWildCard10       254260              4621 ns/op            1168 B/op         30 allocs/op
BenchmarkRegexp1         1830240               659.5 ns/op           424 B/op          8 allocs/op
BenchmarkRegexp5          489993              2379 ns/op             760 B/op         19 allocs/op
BenchmarkRegexp10         272499              4337 ns/op            1168 B/op         30 allocs/op
PASS
ok      github.com/bmf-san/goblin       13.572s
```

![スクリーンショット 2022-11-03 1 18 15](https://user-images.githubusercontent.com/13291041/199543868-e6826ee9-c9b0-443f-b4ee-5c6e14ef66c9.png)


# Impact range
No specification change.

# Operational Requirements
N/A

# Related Issue
N/A

# Supplement
N/A